Move `MsgCommitPubRand` and `MsgAddFinalitySig` into msg.rs
